Add onViewCreated, create a separate CrimeAdapter file, change inflater, add setupUI() method
