+++ app/src/main/java/com/example/criminalintentapp/CrimeAdapter.kt
@@ -1,0 +1,51 @@
+package com.example.criminalintentapp
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import android.widget.Toast
+import androidx.recyclerview.widget.RecyclerView
+
+class CrimeAdapter(var crimes: List<Crime>) :
+    RecyclerView.Adapter<CrimeAdapter.CrimeHolder>() {
+
+    class CrimeHolder(itemView: View) : RecyclerView.ViewHolder(itemView), View.OnClickListener {
+
+        private lateinit var crime: Crime
+
+        private val titleTextView: TextView = itemView.findViewById(R.id.crime_title)
+        private val dateTextView: TextView = itemView.findViewById(R.id.crime_date)
+
+        init {
+            itemView.setOnClickListener(this)
+        }
+
+        fun bind(crime: Crime) {
+            this.crime = crime
+            titleTextView.text = this.crime.title
+            dateTextView.text = this.crime.date.toString()
+        }
+
+        override fun onClick(v: View?) {
+            if (v != null) {
+                Toast.makeText(v.context, "${crime.title} pressed", Toast.LENGTH_SHORT).show()
+            }
+        }
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CrimeHolder {
+        val v: View =
+            LayoutInflater.from(parent.context).inflate(R.layout.list_item_crime, parent, false)
+        return CrimeHolder(v)
+    }
+
+    override fun onBindViewHolder(holder: CrimeHolder, position: Int) {
+        val crime = crimes[position]
+        holder.bind(crime)
+    }
+
+    override fun getItemCount(): Int {
+        return crimes.size
+    }
+}
