+++ app/src/main/java/com/example/criminalintentapp/CrimeListFragment.kt
@@ -2,17 +2,13 @@ package com.example.criminalintentapp
 
 import android.os.Bundle
 import android.util.Log
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 
-class CrimeListFragment : Fragment() {
+class CrimeListFragment : Fragment(R.layout.fragment_crime_list) {
 
     private lateinit var crimeRecyclerView: RecyclerView
     private var adapter: CrimeAdapter? = null
@@ -26,72 +22,20 @@ class CrimeListFragment : Fragment() {
         Log.d(TAG, "Total crimes: ${crimeListViewModel.crimes.size}")
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        val view = inflater.inflate(R.layout.fragment_crime_list, container, false)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
-        bindViews(view)
+        setupUI(view)
+    }
+
+    private fun setupUI(view: View) {
+        crimeRecyclerView = view.findViewById(R.id.crime_recycler_view) as RecyclerView
 
         crimeRecyclerView.layoutManager = LinearLayoutManager(context)
 
-        updateUI()
-
-        return view
-    }
-
-    private fun bindViews(view: View) {
-        crimeRecyclerView = view.findViewById(R.id.crime_recycler_view) as RecyclerView
-    }
-
-    private fun updateUI() {
         val crimes = crimeListViewModel.crimes
         adapter = CrimeAdapter(crimes)
         crimeRecyclerView.adapter = adapter
-    }
-
-    private inner class CrimeHolder(view: View) :
-        RecyclerView.ViewHolder(view),
-        View.OnClickListener {
-
-        private lateinit var crime: Crime
-
-        private val titleTextView: TextView = itemView.findViewById(R.id.crime_title)
-        private val dateTextView: TextView = itemView.findViewById(R.id.crime_date)
-
-        init {
-            itemView.setOnClickListener(this)
-        }
-
-        fun bind(crime: Crime) {
-            this.crime = crime
-            titleTextView.text = this.crime.title
-            dateTextView.text = this.crime.date.toString()
-        }
-
-        override fun onClick(v: View?) {
-            Toast.makeText(context, "${crime.title} pressed", Toast.LENGTH_SHORT).show()
-        }
-    }
-
-    private inner class CrimeAdapter(var crimes: List<Crime>) :
-        RecyclerView.Adapter<CrimeHolder>() {
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CrimeHolder {
-            val view = layoutInflater.inflate(R.layout.list_item_crime, parent, false)
-            return CrimeHolder(view)
-        }
-
-        override fun onBindViewHolder(holder: CrimeHolder, position: Int) {
-            val crime = crimes[position]
-            holder.bind(crime)
-        }
-
-        override fun getItemCount(): Int {
-            return crimes.size
-        }
     }
 
     companion object {
