+++ app/src/main/java/com/example/criminalintentapp/CrimeListViewModel.kt
@@ -7,11 +7,15 @@ class CrimeListViewModel : ViewModel() {
     val crimes = mutableListOf<Crime>()
 
     init {
-        for (i in 0 until 100) {
+        for (i in 0 until NUMBER_OF_ALL_CRIMES) {
             val crime = Crime()
             crime.title = "Crime #$i"
             crime.isSolved = i % 2 == 0
             crimes += crime
         }
+    }
+
+    companion object {
+        private const val NUMBER_OF_ALL_CRIMES = 100
     }
 }
